feat: 구조 구성

+++ src/main/java/com/example/SentiStock_backend/purchase/domain/entity/PurchaseEntity.java
@@ -47,5 +47,5 @@ public class PurchaseEntity {
     private Float avgPrice;
 
     @Column(name = "pur_senti", nullable = false)
-    private Long purSenti;
+    private Double purSenti;
 }
